% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{reqwetdur}
\alias{reqwetdur}
\title{Wet duration requirement for critical disease threshold at a given temperature}
\usage{
reqwetdur(x, wmin, wmax, tmin, tmax, topt)
}
\arguments{
\item{x}{a scalar, vector or matrix of temperatures at which to estimate 
response function}

\item{wmin, wmax}{scalar or vector minimum and maximum wetness duration 
requirements}

\item{tmin, tmax, topt}{scalar or vector minimum, maximum and optimum 
temperatures for infection, respectively; see Details}
}
\value{
A scalar, vector or matrix
}
\description{
Function \code{wetdur} estimates the wetness duration requirement (hours)
for the critical disease threshold for a given temperature and given
minimum, \code{wmin}, and maximum, \code{wmax}, wetness duration requirements.
}
\details{
Calls \code{tempresp}. Note that \code{wmin} and \code{wmax} must be compatible
with and \code{topt} are scalar and \code{nrow(t)} > 1,
then their values are repeated \code{ncol{t}}. \code{tmin}, \code{tmax} and 
\code{topt} must be the same length. If \code{length(tmin)} > 1 then 
\code{length(tmin)} must equal \code{ncol{t}}, or, if \code{ncol{t}} = 1,
then columns are \code{t} are repeated \code{length(tmin)} times.
}
\examples{

# phytophthora cactorum (apple fruit)
reqwetdur(0:40, 2, 5, 1, 35, 25)

}
\references{
Magarey, R. D., T. B. Sutton, and C. L. Thayer. (2005). A simple generic 
infection model for foliar fungal plant pathogens. Phytopathology 95 (1): 
92–100. \doi{10.1094/PHYTO-95-0092}
}
\seealso{
\link{tempresp}
}
