% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{tempresp}
\alias{tempresp}
\title{Temperature response curve for simple generic infection model}
\usage{
tempresp(x, tmin, tmax, topt)
}
\arguments{
\item{x}{a scalar, vector or matrix of temperatures at which to estimate 
response function}

\item{tmin, tmax, topt}{scalar or vector minimum, maximum and optimum 
temperatures for infection, respectively; see Details}
}
\value{
A scalar, vector or matrix
}
\description{
Function \code{tresp} estimates the temperature response function of a disease
given minimum, \code{tmin}, maximum, \code{tmax}, and optimum, \code{topt}, 
temperatures for infection.
}
\details{
If \code{tmin}, \code{tmax} and \code{topt} are scalar and \code{nrow(t)} > 1,
then their values are repeated \code{ncol{t}}. \code{tmin}, \code{tmax} and 
\code{topt} must be the same length. If \code{length(tmin)} > 1 then 
\code{length(tmin)} must equal \code{ncol{t}}, or, if \code{ncol{t}} = 1,
then columns are \code{t} are repeated \code{length(tmin)} times.
}
\examples{

# phytophthora cactorum (apple fruit)
tempresp(0:40, 1, 35, 25)

}
\references{
Magarey, R. D., T. B. Sutton, and C. L. Thayer. (2005). A simple generic 
infection model for foliar fungal plant pathogens. Phytopathology 95 (1): 
92–100. \doi{10.1094/PHYTO-95-0092}
}
